change: rename `qovery connect` to `qovery dev` feat: add -c option to change the root directory of local configuration

+++ api/local_persistence.go
@@ -37,12 +37,24 @@ func SetAccountId(id string) {
 	_ = ioutil.WriteFile(filePath, []byte(id), 0755)
 }
 
-func SaveLocalConfiguration(data map[string]interface{}) {
+func SaveLocalConfiguration(root string, data map[string]interface{}) {
 	j, _ := json.Marshal(data)
-	_ = os.MkdirAll(".qovery", 0755)
-	_ = ioutil.WriteFile(filepath.FromSlash(".qovery/local_configuration.json"), j, 0755)
+
+	if root == "" {
+		_ = os.MkdirAll(".qovery", 0755)
+		_ = ioutil.WriteFile(filepath.FromSlash(".qovery/local_configuration.json"), j, 0755)
+		return
+	}
+
+	_ = os.MkdirAll(filepath.FromSlash(fmt.Sprintf("%s/.qovery", root)), 0755)
+	_ = ioutil.WriteFile(filepath.FromSlash(fmt.Sprintf("%s/.qovery/local_configuration.json", root)), j, 0755)
 }
 
-func DeleteLocalConfiguration() {
-	_ = os.Remove(filepath.FromSlash(".qovery/local_configuration.json"))
+func DeleteLocalConfiguration(root string) {
+	if root == "" {
+		_ = os.Remove(filepath.FromSlash(".qovery/local_configuration.json"))
+		return
+	}
+
+	_ = os.Remove(filepath.FromSlash(fmt.Sprintf("%s/.qovery/local_configuration.json", root)))
 }
